Add generics to some utility functions

+++ sprec/util_test.go
@@ -13,33 +13,33 @@ import (
 var _ = Describe("Util", func() {
 
 	Specify("Abs", func() {
-		Expect(Abs(-0.1)).To(EqualFloat32(0.1))
-		Expect(Abs(-13.57)).To(EqualFloat32(13.57))
-		Expect(Abs(11.01)).To(EqualFloat32(11.01))
+		Expect(Abs(float32(-0.1))).To(EqualFloat32(0.1))
+		Expect(Abs(float32(-13.57))).To(EqualFloat32(13.57))
+		Expect(Abs(float32(11.01))).To(EqualFloat32(11.01))
 	})
 
 	Specify("Max", func() {
-		Expect(Max(1.0, 2.0)).To(EqualFloat32(2.0))
-		Expect(Max(1.0, -1.0)).To(EqualFloat32(1.0))
-		Expect(Max(5.0, 5.0)).To(EqualFloat32(5.0))
+		Expect(Max(float32(1.0), float32(2.0))).To(EqualFloat32(2.0))
+		Expect(Max(float32(1.0), float32(-1.0))).To(EqualFloat32(1.0))
+		Expect(Max(float32(5.0), float32(5.0))).To(EqualFloat32(5.0))
 	})
 
 	Specify("Min", func() {
-		Expect(Min(1.0, 2.0)).To(EqualFloat32(1.0))
-		Expect(Min(1.0, -1.0)).To(EqualFloat32(-1.0))
-		Expect(Min(5.0, 5.0)).To(EqualFloat32(5.0))
+		Expect(Min(float32(1.0), float32(2.0))).To(EqualFloat32(1.0))
+		Expect(Min(float32(1.0), float32(-1.0))).To(EqualFloat32(-1.0))
+		Expect(Min(float32(5.0), float32(5.0))).To(EqualFloat32(5.0))
 	})
 
 	Specify("Clamp", func() {
-		Expect(Clamp(1.0, 2.0, 3.0)).To(EqualFloat32(2.0))
-		Expect(Clamp(2.5, 2.0, 3.0)).To(EqualFloat32(2.5))
-		Expect(Clamp(4.0, 2.0, 3.0)).To(EqualFloat32(3.0))
+		Expect(Clamp(float32(1.0), float32(2.0), float32(3.0))).To(EqualFloat32(2.0))
+		Expect(Clamp(float32(2.5), float32(2.0), float32(3.0))).To(EqualFloat32(2.5))
+		Expect(Clamp(float32(4.0), float32(2.0), float32(3.0))).To(EqualFloat32(3.0))
 	})
 
 	Specify("Mix", func() {
-		Expect(Mix(1.0, 2.0, 0.0)).To(EqualFloat32(1.0))
-		Expect(Mix(1.0, 2.0, 1.0)).To(EqualFloat32(2.0))
-		Expect(Mix(1.0, 2.0, 0.5)).To(EqualFloat32(1.5))
+		Expect(Mix(float32(1.0), float32(2.0), float32(0.0))).To(EqualFloat32(1.0))
+		Expect(Mix(float32(1.0), float32(2.0), float32(1.0))).To(EqualFloat32(2.0))
+		Expect(Mix(float32(1.0), float32(2.0), float32(0.5))).To(EqualFloat32(1.5))
 	})
 
 	Specify("Eq", func() {
